decouple IPC folder operation
